[ty] Refactor to use contains_identifier
